requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.32.0
motor==3.7.1
pymongo[srv]==4.14.1
pydantic[email]==2.11.7
python-dotenv==1.1.1
python-multipart==0.0.9

passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

numpy==2.3.2
Pillow==11.3.0
torch==2.2.2
facenet-pytorch==2.6.0
